LimiterHandler: Replacing decimal with float

+++ holmes/handlers/limiter.py
@@ -3,7 +3,6 @@
 
 from ujson import loads
 from tornado.gen import coroutine
-from decimal import Decimal
 
 from holmes.models import Limiter, User
 from holmes.handlers import BaseHandler
@@ -23,7 +22,7 @@ class LimiterHandler(BaseHandler):
 
             percentage = 0
             if limit.value > 0:
-                percentage = Decimal(current_value) / limit.value
+                percentage = float(current_value) / limit.value
 
             result.append({
                 'id': limit.id,
